requirements.txt
# ======= Core Flask Web App & Configuration =======
flask==3.0.3
python-dotenv==1.0.1
gunicorn

# ======= LangChain Framework (Stable Versions) =======
langchain~=0.2.5
langchain-core~=0.2.10
langchain-community~=0.2.5
langchain-astradb~=0.4.0
langchain-groq~=0.1.5
sentence-transformers~=2.7.0

# ======= Data Handling & Document Loading =======
pandas==2.2.2
pypdf==4.2.0
datasets==2.19.1

# ======= Monitoring =======
prometheus_client==0.20.0